LinkedIn Ad Library API Reference

This file provides comprehensive documentation and reference information for the LinkedIn Ad Library API, based on the official documentation.

API Overview

LinkedIn's Ad Library application programming interface (API) helps you to perform custom searches on ad data stored in the platform based on advertiser, keywords, served dates and served countries.

Authentication

The API uses OAuth 2.0 authentication. You must:

Configure your LinkedIn app credentials using li_auth_configure()

Authenticate using li_auth() to obtain an access token

Use li_query() which automatically uses the cached token

Query Parameters

The API accepts the following query parameters:

keyword

String. Keyword that will be searched in the content of the ad. If there are multiple keywords, separate them using a space. Multiple keywords are treated as a logical AND operation.

countries

Array of CountryUrns. CountryUrn is URN representation of lowercase 2-letter ISO country codes.

advertiser

String. Advertiser name that will be searched against the advertiser of the ad.

dateRange

DateRange. Date range to filter ads on the served dates. Start represents the inclusive start time, End represents the exclusive end time.

start

Integer. Used for pagination. Starting index of results.

count

Integer. Used for pagination. Maximum 25 ads are returned and count should be less than 25.

Response Schema

The API returns an AdsCreativeTransparencyEntity with the following structure:

adUrl

URL pointing to ads preview URL

details

AdTransparencyCreativeEntityDetails object containing:

advertiser: AdvertiserDetails (name, URL, payer)

type: Ad format type (e.g., "SPOTLIGHT_V2", "TEXT_AD")

adTargeting: Array of targeting parameters

adStatistics: Impression statistics and date ranges

isRestricted

Boolean indicating if ad violates LinkedIn's Advertising Policy

restrictionDetails

String with restriction information (restricted ads only)

Example Usage

# First-time setup
li_auth_configure()  # Enter your LinkedIn app credentials
li_auth()           # Authenticate via browser

# Search for ads by keyword and country
ads_by_keyword <- li_query(
  keyword = "data science",
  countries = c("us", "ca"),
  start_date = "2024-01-01",
  end_date = "2024-03-31"
)

# Search for ads by advertiser
ads_by_advertiser <- li_query(
  advertiser = "Google",
  countries = c("gb"),
  start_date = "2024-05-01",
  end_date = "2024-06-01",
  max_pages = 2
)

Rate Limits and Best Practices

Maximum 25 results per request

Use pagination with start and count parameters

Date ranges: start is inclusive, end is exclusive

Country codes should be lowercase 2-letter ISO codes

Keywords are searched with logical AND operation

Data Structure Details

The returned data frame contains the following columns:

ad_url

Direct link to the ad preview

is_restricted

Boolean indicating policy violations

restriction_details

Details about restrictions (if any)

advertiser_name

Name of the advertising organization

advertiser_url

LinkedIn page URL of the advertiser

ad_payer

Entity that paid for the ad

ad_type

Format/type of the advertisement

first_impression_at

First time the ad was shown (POSIXct)

latest_impression_at

Most recent time the ad was shown (POSIXct)

total_impressions_from

Lower bound of total impressions

total_impressions_to

Upper bound of total impressions

impressions_by_country

List-column with country-wise impression data

ad_targeting

List-column with targeting criteria details

References

LinkedIn Ad Library API Documentation: https://www.linkedin.com/ad-library/api/ads